refactor: share non-packed table dispatch
