+++ crates/evm2/src/interpreter/dispatch/table/mod.rs
@@ -1,6 +1,11 @@
 use super::{InspectMode, UNKNOWN_OP, inc_pc, run_state, unknown_instruction};
 #[cfg(dispatch_packed)]
 use crate::interpreter::gas::RemainingGas;
+#[cfg(not(dispatch_packed))]
+use crate::{
+    EvmConfig,
+    interpreter::{InterpreterState, Pc, Result, Stack, private::InstructionImplFn},
+};
 use crate::{
     EvmTypes,
     interpreter::{InstrStop, Interpreter, gas::Gas},
@@ -50,6 +55,48 @@ const fn finish_loop(gas: &mut Gas, remaining_gas: LoopState) {
 #[inline(always)]
 #[cfg(not(dispatch_packed))]
 const fn finish_loop(_gas: &mut Gas, _loop_state: LoopState) {}
+
+#[cold] // Not cold, but avoids MIR inlining.
+#[inline(always)]
+#[cfg(not(dispatch_packed))]
+fn dispatch_mono<T: EvmTypes, C: EvmConfig<T>, M: InspectMode<T>, const OP: u8>(
+    mut pc: Pc,
+    mut stack: Stack<'_>,
+    state: &mut InterpreterState<'_, T>,
+    instr: InstructionImplFn<T>,
+) -> (Pc, usize) {
+    if M::INSPECT {
+        M::step(state, pc, stack.len);
+    }
+    let r;
+    match pre_step::<T, C, OP>(state.gas_mut()) {
+        Ok(()) => {
+            r = instr(&mut pc, stack.as_mut(), state);
+            if !M::INSPECT || r.is_ok() {
+                inc_pc(&mut pc, OP);
+            }
+        }
+        Err(e) => r = Err(e),
+    }
+    if M::INSPECT {
+        state.set_result(r);
+        M::step_end(state, pc, stack.len);
+    }
+    if r.is_err() {
+        cold_path();
+        if !M::INSPECT {
+            state.set_result(r);
+        }
+        return (Pc::new(core::ptr::null()), stack.len);
+    }
+    (pc, stack.len)
+}
+
+#[inline(always)]
+#[cfg(not(dispatch_packed))]
+const fn pre_step<T: EvmTypes, C: EvmConfig<T>, const OP: u8>(gas: &mut Gas) -> Result {
+    gas.spend(C::VERSION_TABLES.static_gas(OP) as _)
+}
 
 pub(in crate::interpreter) fn run<T: EvmTypes>(
     interpreter: &mut Interpreter<'_, T>,
