+++ crates/evm2/src/interpreter/dispatch/table/unpacked.rs
@@ -1,9 +1,8 @@
 use super::InspectMode;
 use crate::{
     EvmConfig, EvmTypes,
-    interpreter::{InterpreterState, Pc, Result, Stack, gas::Gas, private::InstructionImplFn},
+    interpreter::{InterpreterState, Pc, Stack},
 };
-use core::hint::cold_path;
 
 /// Unpacked instruction return value.
 type InstrFnRet = (*const u8, usize);
@@ -37,12 +36,13 @@ extern_table! {
         state: &mut InterpreterState<'_, T>,
     ) -> InstrFnRet {
         let _ = DYNAMIC_GAS;
-        dispatch_mono::<T, C, M, OP>(
+        let (pc, stack_len) = super::dispatch_mono::<T, C, M, OP>(
             pc,
             stack,
             state,
             C::VERSION_TABLES.instruction(OP).instr,
-        )
+        );
+        (pc.as_ptr(), stack_len)
     }
 
     pub(in crate::interpreter::dispatch) fn unknown_dispatch<
@@ -54,51 +54,12 @@ extern_table! {
         stack: Stack<'_>,
         state: &mut InterpreterState<'_, T>,
     ) -> InstrFnRet {
-        dispatch_mono::<T, C, M, { super::UNKNOWN_OP }>(
+        let (pc, stack_len) = super::dispatch_mono::<T, C, M, { super::UNKNOWN_OP }>(
             pc,
             stack,
             state,
             super::unknown_instruction,
-        )
+        );
+        (pc.as_ptr(), stack_len)
     }
-}
-
-#[cold] // Not cold, but avoids MIR inlining.
-#[inline(always)]
-fn dispatch_mono<T: EvmTypes, C: EvmConfig<T>, M: InspectMode<T>, const OP: u8>(
-    mut pc: Pc,
-    mut stack: Stack<'_>,
-    state: &mut InterpreterState<'_, T>,
-    instr: InstructionImplFn<T>,
-) -> InstrFnRet {
-    if M::INSPECT {
-        M::step(state, pc, stack.len);
-    }
-    let r;
-    match pre_step::<T, C, OP>(state.gas_mut()) {
-        Ok(()) => {
-            r = instr(&mut pc, stack.as_mut(), state);
-            if !M::INSPECT || r.is_ok() {
-                super::inc_pc(&mut pc, OP);
-            }
-        }
-        Err(e) => r = Err(e),
-    }
-    if M::INSPECT {
-        state.set_result(r);
-        M::step_end(state, pc, stack.len);
-    }
-    if r.is_err() {
-        cold_path();
-        if !M::INSPECT {
-            state.set_result(r);
-        }
-        return (core::ptr::null(), stack.len);
-    }
-    (pc.as_ptr(), stack.len)
-}
-
-#[inline(always)]
-const fn pre_step<T: EvmTypes, C: EvmConfig<T>, const OP: u8>(gas: &mut Gas) -> Result {
-    gas.spend(C::VERSION_TABLES.static_gas(OP) as _)
 }
